Struct ffmpeg_sys_next::AVBufferSrcParameters

#[repr(C)]
pub struct AVBufferSrcParameters {
    pub format: c_int,
    pub time_base: AVRational,
    pub width: c_int,
    pub height: c_int,
    pub sample_aspect_ratio: AVRational,
    pub frame_rate: AVRational,
    pub hw_frames_ctx: *mut AVBufferRef,
    pub sample_rate: c_int,
    pub channel_layout: u64,
    pub ch_layout: AVChannelLayout,
}

This structure contains the parameters describing the frames that will be passed to this filter.

It should be allocated with av_buffersrc_parameters_alloc() and freed with av_free(). All the allocated fields in it remain owned by the caller.

Fields
format: c_int

video: the pixel format, value corresponds to enum AVPixelFormat audio: the sample format, value corresponds to enum AVSampleFormat

time_base: AVRational

The timebase to be used for the timestamps on the input frames.

width: c_int

Video only, the display dimensions of the input frames.

height: c_int

Video only, the display dimensions of the input frames.

sample_aspect_ratio: AVRational

Video only, the sample (pixel) aspect ratio.

frame_rate: AVRational

Video only, the frame rate of the input video. This field must only be set to a non-zero value if input stream has a known constant framerate and should be left at its initial value if the framerate is variable or unknown.

hw_frames_ctx: *mut AVBufferRef

Video with a hwaccel pixel format only. This should be a reference to an AVHWFramesContext instance describing the input frames.

sample_rate: c_int

Audio only, the audio sampling rate in samples per second.

channel_layout: u64

Audio only, the audio channel layout @deprecated use ch_layout

ch_layout: AVChannelLayout

Audio only, the audio channel layout